requirements.txt
python-dotenv==1.0.1
openai==1.68.2
flask==3.1.0
